requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.4
Flask-Sockets==0.2.1
gevent==21.1.2
gevent-websocket==0.10.1
google-api-core==1.27.0
google-auth==1.30.0
google-cloud-speech==2.3.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.37.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.19
MarkupSafe==2.0.0
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
websockets==9.0.2
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0
